Create an internal API scheme
* which can be used to access defaulting and validation functions without having
to import webhook package
