+++ pkg/internal/BUILD.bazel
@@ -9,7 +9,7 @@ filegroup(
     name = "all-srcs",
     srcs = [
         ":package-srcs",
-        "//pkg/internal/api/validation:all-srcs",
+        "//pkg/internal/api:all-srcs",
         "//pkg/internal/apis/acme:all-srcs",
         "//pkg/internal/apis/certmanager:all-srcs",
         "//pkg/internal/apis/meta:all-srcs",
