+++ pkg/internal/api/BUILD.bazel
@@ -1,0 +1,32 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["scheme.go"],
+    importpath = "github.com/jetstack/cert-manager/pkg/internal/api",
+    visibility = ["//pkg:__subpackages__"],
+    deps = [
+        "//pkg/internal/api/validation:go_default_library",
+        "//pkg/internal/apis/acme/install:go_default_library",
+        "//pkg/internal/apis/certmanager/install:go_default_library",
+        "//pkg/internal/apis/meta/install:go_default_library",
+        "@io_k8s_apimachinery//pkg/runtime:go_default_library",
+    ],
+)
+
+filegroup(
+    name = "package-srcs",
+    srcs = glob(["**"]),
+    tags = ["automanaged"],
+    visibility = ["//visibility:private"],
+)
+
+filegroup(
+    name = "all-srcs",
+    srcs = [
+        ":package-srcs",
+        "//pkg/internal/api/validation:all-srcs",
+    ],
+    tags = ["automanaged"],
+    visibility = ["//visibility:public"],
+)
