+++ pkg/webhook/BUILD.bazel
@@ -1,19 +1,5 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
-go_library(
-    name = "go_default_library",
-    srcs = ["scheme.go"],
-    importpath = "github.com/jetstack/cert-manager/pkg/webhook",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/internal/api/validation:go_default_library",
-        "//pkg/internal/apis/acme/install:go_default_library",
-        "//pkg/internal/apis/certmanager/install:go_default_library",
-        "//pkg/internal/apis/meta/install:go_default_library",
-        "@io_k8s_apimachinery//pkg/runtime:go_default_library",
-    ],
-)
-
 filegroup(
     name = "package-srcs",
     srcs = glob(["**"]),
@@ -31,4 +17,12 @@ filegroup(
     ],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
+)
+
+go_library(
+    name = "go_default_library",
+    srcs = ["scheme.go"],
+    importpath = "github.com/jetstack/cert-manager/pkg/webhook",
+    visibility = ["//visibility:public"],
+    deps = ["//pkg/internal/api:go_default_library"],
 )
