+++ pkg/webhook/scheme.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2019 The Jetstack cert-manager contributors.
+Copyright 2020 The Jetstack cert-manager contributors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -16,37 +16,11 @@ limitations under the License.
 
 package webhook
 
-import (
-	"k8s.io/apimachinery/pkg/runtime"
+import internalcmapi "github.com/jetstack/cert-manager/pkg/internal/api"
 
-	"github.com/jetstack/cert-manager/pkg/internal/api/validation"
-	acmeinstall "github.com/jetstack/cert-manager/pkg/internal/apis/acme/install"
-	cminstall "github.com/jetstack/cert-manager/pkg/internal/apis/certmanager/install"
-	metainstall "github.com/jetstack/cert-manager/pkg/internal/apis/meta/install"
-)
-
-// Define a Scheme that has all cert-manager API types registered, including
-// the internal API version, defaulting functions and conversion functions for
-// all external versions.
-// This scheme should *only* be used by the webhook as the conversion/defaulter
-// functions are likely to change in future, and all controllers consuming
-// cert-manager APIs should have a consistent view of all API kinds.
-
+// These public aliases are required because the ./cmd/webhook packages can't import them from ./pkg/internal
+// TODO: Create a top-level ./internal package which can be used by ./cmd and ./test
 var (
-	// Scheme is a Kubernetes runtime.Scheme with all internal and external API
-	// versions for cert-manager types registered.
-	Scheme = runtime.NewScheme()
-
-	// ValidationRegistry is a validation registry with all required
-	// validations that should be enforced by the webhook component.
-	ValidationRegistry = validation.NewRegistry(Scheme)
+	Scheme             = internalcmapi.Scheme
+	ValidationRegistry = internalcmapi.ValidationRegistry
 )
-
-func init() {
-	cminstall.Install(Scheme)
-	acmeinstall.Install(Scheme)
-	metainstall.Install(Scheme)
-
-	cminstall.InstallValidation(ValidationRegistry)
-	acmeinstall.InstallValidation(ValidationRegistry)
-}
